Fixes and style improvements

+++ src/Controllers/CategoryController.php
@@ -17,6 +17,8 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', Category::class);
+
         $categories = Category::all();
         return view('laralum_blog::categories.index', ['categories' => $categories]);
     }
@@ -28,6 +30,8 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Category::class);
+
         return view('laralum_blog::categories.create');
     }
 
@@ -39,9 +43,10 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Category::class);
+
         $this->validate($request, [
-            'title' => 'required|min:5|max:50',
-            'description' => 'required|max:100',
+            'name' => 'required|min:5|max:50',
         ]);
         Category::create($request->all());
         return redirect()->route('laralum::blog.categories.index')->with('success', __('laralum_blog::category_added'));
@@ -55,6 +60,8 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
+        $this->authorize('view', Category::class);
+
         return view('laralum_blog::categories.show', ['category' => $category]);
     }
 
@@ -66,6 +73,8 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $this->authorize('update', $category);
+
         return view('laralum_blog::categories.edit', ['category' => $category]);
     }
 
@@ -78,9 +87,10 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('update', $category);
+
         $this->validate($request, [
-            'title' => 'required|min:5|max:50',
-            'description' => 'required|max:100',
+            'name' => 'required|min:5|max:50',
         ]);
         $category->update($request->all());
         return redirect()->route('laralum::blog.categories.index')->with('success', __('laralum_blog::category_updated',['id' => $category->id]));
@@ -94,10 +104,11 @@ class CategoryController extends Controller
      */
     public function confirmDestroy(Category $category)
     {
+        $this->authorize('delete', $category);
 
         return view('laralum::pages.confirmation', [
             'method' => 'DELETE',
-            'message' => __('laralum_blog::general.sure_del_category', ['category' => $category->title]),
+            'message' => __('laralum_blog::general.sure_del_category', ['category' => $category->name]),
             'action' => route('laralum::blog.categories.destroy', ['category' => $category->id]),
         ]);
     }
@@ -110,6 +121,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+
         $category->deleteComments();
         $category->deletePosts();
         $category->delete();
